requirements.txt
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
cymem==2.0.6
fastai==2.5.3
fastcore==1.3.27
fastdownload==0.0.5
fastprogress==1.0.2
fonttools==4.29.1
idna==3.3
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.0
matplotlib==3.5.1
murmurhash==1.0.6
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pathy==0.6.1
Pillow==9.0.1
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
smart-open==5.2.1
spacy==3.2.2
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.1.0
timm==0.5.4
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tqdm==4.62.3
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.8
wasabi==0.9.0
